Poll session of live flight search.

Returns live prices from all our suppliers for the requested flight itinerary. This function is GET step. Make sure you have set API using apiSetKey before.

See https://rapidapi.com/skyscanner/api/skyscanner-flight-search.

apiPollSession(response, sortType = c("price", "duration", "carrier",
  "outboundarrivetime", "outbounddeparttime", "inboundarrivetime",
  "inbounddeparttime"), sortOrder = c("asc", "desc"), duration = NULL,
  stops = NULL, includeCarriers = NULL, excludeCarriers = NULL,
  originAirports = NULL, destinationAirports = NULL,
  outboundDepartTime = NULL, outboundDepartStartTime = NULL,
  outboundDepartEndTime = NULL, outboundArriveStartTime = NULL,
  outboundArriveEndTime = NULL, inboundDepartTime = NULL,
  inboundDepartStartTime = NULL, inboundDepartEndTime = NULL,
  inboundArriveStartTime = NULL, inboundArriveEndTime = NULL)

Arguments

response
Return object of apiCreateSession.

sortType
(OPTIONAL) The parameter to sort results on. Can be "price", "duration", "carrier", "outboundarrivetime", "outbounddeparttime", "inboundarrivetime", "inbounddeparttime".

sortOrder
(OPTIONAL) The sort order. "asc" or "desc".

duration
(OPTIONAL) Filter for maximum duration in minutes. Integer between 0 and 1800.

stops
(OPTIONAL) Filter by number of stops. 0: direct flights only. 1: flights with one stop only. To show all flights do not use (only supports values 0 and 1).

includeCarriers
(OPTIONAL) Filter flights by the specified carriers. Must be semicolon-separated IATA codes.

excludeCarriers
(OPTIONAL) Filter flights by any but the specified carriers. Must be semicolon-separated IATA codes.

originAirports
(OPTIONAL) Origin airports to filter on. List of airport codes delimited by ';'.

destinationAirports
(OPTIONAL) Destination airports to filter on. List of airport codes delimited by ';'.

outboundDepartTime
(OPTIONAL) Filter for outbound departure time by time period of the day (i.e. morning, afternoon, evening). List of day time period delimited by ';' (acceptable values are "M", "A", "E").

outboundDepartStartTime
(OPTIONAL) Filter for start of range for outbound departure time. Format "hh:mm".

outboundDepartEndTime
(OPTIONAL) Filter for end of range for outbound departure time. Format "hh:mm".

outboundArriveStartTime
(OPTIONAL) Filter for start of range for outbound arrival time. Format "hh:mm".

outboundArriveEndTime
(OPTIONAL) Filter for end of range for outbound arrival time. Format "hh:mm".

inboundDepartTime
(OPTIONAL) Filter for inbound departure time by time period of the day (i.e. morning, afternoon, evening). List of day time period delimited by ';' (acceptable values are "M", "A", "E").

inboundDepartStartTime
(OPTIONAL) Filter for start of range for inbound departure time. Format "hh:mm".

inboundDepartEndTime
(OPTIONAL) Filter for end of range for inbound departure time. Format "hh:mm".

inboundArriveStartTime
(OPTIONAL) Filter for start of range for inbound arrival time. Format "hh:mm".

inboundArriveEndTime
(OPTIONAL) Filter for end of range for inbound arrival time. Format "hh:mm".

Value

A response() object of request.

See also

Examples

# NOT RUN {
apiSetKey("YOUR_API_KEY")
resp <- apiCreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
apiPollSession(resp)
# }